feat(structure): reorganiza projeto no padrão v5 (src/, sql_scripts/, blueprints, bootstrap Flask)
