+++ sql_scripts/legacy/02_create_functions.sql
